pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["bogdankozlovsky <bogdan.k@datarails.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
Pillow = "^9.0.1"
Statis = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
